requirements.txt
black==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
flake8==7.0.0
idna==3.7
ImageHash==4.3.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyGObject==3.48.1
pytest==8.1.1
python-dotenv==1.0.1
PyWavelets==1.6.0
requests==2.31.0
scipy==1.13.1
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
websockets==12.0
